Minor ui and functionality changes

+++ passpro/src/App.jsx
@@ -1,6 +1,9 @@
 import { useState, useRef, useEffect } from "react";
 import { v4 as uuidv4 } from "uuid";
 import "./App.css";
+import logo from "./assets/LOGO2.png";
+import eyeo from "./assets/eyeo.svg";
+import eyec from "./assets/eyec.svg";
 
 function App() {
   //localStorage.clear();
@@ -20,23 +23,21 @@ function App() {
   useEffect(() => {
     getPass();
   }, []);
+
+  const [visible, setVisible] = useState(true);
   const changeVis = () => {
-    console.log("fn called");
-    const s = ref.current.src;
-    console.log(s);
-    if (s === "https://cdn-icons-png.flaticon.com/128/10969/10969091.png") {
-      console.log("pv");
-      alert("Password is Visible");
-      ref.current.src =
-        "https://cdn-icons-png.flaticon.com/128/10968/10968680.png";
-    } else if (
-      s === "https://cdn-icons-png.flaticon.com/128/10968/10968680.png"
-    ) {
-      console.log("cv");
-      alert("Password is hidden");
-      ref.current.src =
-        "https://cdn-icons-png.flaticon.com/128/10969/10969091.png";
+    if (visible) {
+      const prompt = window.confirm("Do you want to hide password?");
+      if (prompt) {
+        ref.current.src = eyec;
+      }
+    } else {
+      const prompt = window.confirm("Do you want to show password?");
+      if (prompt) {
+        ref.current.src = eyeo;
+      }
     }
+    setVisible(!visible);
   };
 
   const handleChange = (e) => {
@@ -45,7 +46,7 @@ function App() {
 
   const storePass = async () => {
     console.log(form);
-    if (form != { url: "", username: "", password: "" }) {
+    if (form.url !== "" && form.username !== "" && form.password !== "") {
       setPasswordArray([...passwordArray, { ...form, id: uuidv4() }]);
       await fetch("http://localhost:3000/", {
         method: "DELETE",
@@ -62,6 +63,8 @@ function App() {
         JSON.stringify([...passwordArray, form])
       );*/
       console.log(passwordArray);
+    } else {
+      alert("Please fill all the fields");
     }
     useForm({ url: "", username: "", password: "" });
   };
@@ -78,18 +81,13 @@ function App() {
   };
   return (
     <>
-      <nav className="bg-rgb(234, 255, 234) shadow-sm h-20 flex items-center justify-center gap-230">
-        <div className="flex items-center justify-center gap-1">
+      <nav className="bg-rgb(253, 215, 203) shadow-sm h-20 flex items-center justify-center gap-220">
+        <div className="flex items-center justify-center gap-3">
           <a href="/">
-            <img
-              src="https://cdn-icons-png.freepik.com/256/14103/14103827.png?uid=R204064089&ga=GA1.1.37188644.1740345148&semt=ais_hybrid"
-              className="size-15"
-            ></img>
+            <img src={logo} alt="Logo" className="size-15 rounded-sm"></img>
           </a>
           <a href="/">
-            <h1 className="text-green-600 font-semibold text-4xl mt-2.5">
-              PassMtrX
-            </h1>
+            <h1 className="text-orange-600 font-semibold text-4xl">PassMtrX</h1>
           </a>
         </div>
         <button className="rounded-2xl">
@@ -101,7 +99,7 @@ function App() {
           </a>
         </button>
       </nav>
-      <h1 className="font-medium text-green-400 text-4xl justify-self-center mt-12">
+      <h1 className="font-medium text-orange-600 text-4xl justify-self-center mt-12">
         Secure Your Passwords in a single MatriX!
       </h1>
       <div className="flex-col justify-self-center justify-center w-1/2 mt-8 h-50">
@@ -110,7 +108,7 @@ function App() {
           id=""
           type="text"
           placeholder="Enter website URL"
-          className="border-2 border-green-400 rounded-4xl w-1/1 h-1/4 mb-4 p-4 focus:border-green-600 focus:outline focus:outline-green-600"
+          className="border-2 border-orange-600 rounded-4xl w-1/1 h-1/4 mb-4 p-4 focus:border-orange-600 focus:outline focus:outline-orange-600"
           value={form.url}
           onChange={handleChange}
           name="url"
@@ -120,7 +118,7 @@ function App() {
           id=""
           type="text"
           placeholder="Enter Username"
-          className="border-2 border-green-400 rounded-4xl w-1/1 h-1/4 mb-4 p-4 focus:border-green-600 focus:outline focus:outline-green-600"
+          className="border-2 border-orange-600 rounded-4xl w-1/1 h-1/4 mb-4 p-4 focus:border-orange-600 focus:outline focus:outline-orange-600"
           value={form.username}
           onChange={handleChange}
           name="username"
@@ -131,26 +129,28 @@ function App() {
             id=""
             type="text"
             placeholder="Enter Password"
-            className="border-2 border-green-400 rounded-4xl w-1/1 h-1/4 p-4 mb-6 focus:border-green-600 focus:outline focus:outline-green-600"
+            className="border-2 border-orange-600 rounded-4xl w-1/1 h-1/4 p-4 mb-6 focus:border-orange-600 focus:outline focus:outline-orange-600"
             value={form.password}
             name="password"
             onChange={handleChange}
           ></input>
-          <img
-            ref={ref}
-            src="https://cdn-icons-png.flaticon.com/128/10969/10969091.png"
-            className="absolute size-8 ml-180"
-            onClick={changeVis}
-          ></img>
+          <button className="cursor-pointer">
+            <img
+              ref={ref}
+              src={eyeo}
+              className="absolute size-8 top-3.5 left-180"
+              onClick={changeVis}
+            ></img>
+          </button>
         </div>
         <button
-          className="rounded-4xl bg-green-400 p-2 w-1/8 hover:bg-green-500 ml-82"
+          className="rounded-4xl bg-orange-600 p-2 w-1/8 hover:bg-orange-700 text-white ml-82 cursor-pointer"
           onClick={storePass}
         >
           Save
         </button>
       </div>
-      <h2 className="text-green-400 font-medium mt-24 justify-self-center text-2xl">
+      <h2 className="text-orange-600 font-medium mt-24 justify-self-center text-2xl">
         Your Passwords
       </h2>
       {passwordArray.length === 0 && (
@@ -158,7 +158,7 @@ function App() {
       )}
       {passwordArray.length != 0 && (
         <table class="table-auto justify-self-center mt-6 w-1/2 rounded-md overflow-hidden mb-8">
-          <thead className="bg-green-600 text-white">
+          <thead className="bg-orange-700 text-white">
             <tr>
               <th className="text-center py-2">Website</th>
               <th className="text-center py-2">Username</th>
@@ -166,7 +166,7 @@ function App() {
               <th className="text-center py-2">Actions</th>
             </tr>
           </thead>
-          <tbody className="bg-green-200">
+          <tbody className="bg-orange-200">
             {passwordArray.map((item, index) => {
               return (
                 <tr key={index}>
@@ -179,7 +179,7 @@ function App() {
                   <td className="text-center w-52 py-1">{item.password}</td>
                   <td className="flex justify-center p-2">
                     <button
-                      className="size-7"
+                      className="size-7 cursor-pointer"
                       onClick={() => deletePass(index)}
                     >
                       <img src="https://cdn-icons-png.flaticon.com/128/10995/10995755.png"></img>
